cortex_eval/requirements.txt
aiohttp>=3.8.0
nltk>=3.8.0
PyYAML>=6.0
rouge>=1.0.1
sentence-transformers>=2.2.0
tqdm>=4.65.0